[sally4405/#4] Refactor: 통합테스트명 변경
